requirements.txt
requests>=2.32.0
openai>=1.0.0
PyYAML>=6.0.1
playwright>=1.42.0
